+++ packages/openapi3/test/output-file.test.ts
@@ -134,4 +134,60 @@ describe("openapi3: output file", () => {
       });
     });
   });
+
+  describe("Predefined variable name behavior", () => {
+    interface ServiceNameCase {
+      description: string;
+      code: string;
+      outputFilePattern: string;
+      expectedOutputFiles: string[];
+    }
+    it.each([
+      // {service-name} cases
+      {
+        description: "{service-name} for one service",
+        code: "@service namespace AAA { model M {a: string} }",
+        outputFilePattern: "{service-name}.yaml",
+        expectedOutputFiles: ["AAA.yaml"],
+      },
+      {
+        description: "{service-name} for multiple services",
+        code:
+          "@service namespace AAA { model M {a: string} }" +
+          "@service namespace BBB { model N {b: string} }",
+        outputFilePattern: "{service-name}.yaml",
+        expectedOutputFiles: ["AAA.yaml", "BBB.yaml"],
+      },
+      // {service-name-if-multiple} cases
+      {
+        description: "{service-name-if-multiple} for one service",
+        code: "@service namespace AAA { model M {a: string} }",
+        outputFilePattern: "{service-name-if-multiple}.yaml",
+        expectedOutputFiles: ["yaml"],
+      },
+      {
+        description: "{service-name-if-multiple} for multiple services",
+        code:
+          "@service namespace AAA { model M {a: string} }" +
+          "@service namespace BBB { model N {b: string} }",
+        outputFilePattern: "{service-name-if-multiple}.yaml",
+        expectedOutputFiles: ["AAA.yaml", "BBB.yaml"],
+      },
+      // fixed name cases
+      {
+        description: "fixed name for one service",
+        code: "@service namespace AAA { model M {a: string} }",
+        outputFilePattern: "fixed-name.yaml",
+        expectedOutputFiles: ["fixed-name.yaml"],
+      },
+    ])("$description", async (c: ServiceNameCase) => {
+      await compileOpenAPI(
+        {
+          "output-file": c.outputFilePattern,
+        },
+        c.code,
+      );
+      for (const outputFile of c.expectedOutputFiles) expectHasOutput(outputFile);
+    });
+  });
 });
